requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
Django==2.2.24
djangorestframework==3.11.2
idna==3.3
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7